Added new feature: While deleting lekarz from karta pobytu all his Recepty, Operacje and Diagnozy are deleted too

+++ src/main/java/com/szbd/hospital/dao/KartaPobytuDAOImpl.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
+import java.util.ArrayList;
 import java.util.List;
 
 @Repository
@@ -71,9 +72,9 @@ public class KartaPobytuDAOImpl implements KartaPobytuDAO {
         KartaPobytu kartaPobytu = entityManager.find(KartaPobytu.class, id_karty);
         Lekarz lekarz = entityManager.find(Lekarz.class, id_lekarza);
 
-        if(kartaPobytu != null && lekarz != null){
-            for (Lekarz lekKarta: kartaPobytu.getLekarze()){
-                if(lekKarta.getId_lekarza() == id_lekarza){
+        if (kartaPobytu != null && lekarz != null) {
+            for (Lekarz lekKarta : kartaPobytu.getLekarze()) {
+                if (lekKarta.getId_lekarza() == id_lekarza) {
                     return;
                 }
             }
@@ -94,36 +95,40 @@ public class KartaPobytuDAOImpl implements KartaPobytuDAO {
 
     @Override
     public void deleteLekarzFromKarta(int id_karty, int id_lekarza) {
+        System.out.println("id_karty: " + id_karty + " id_lekarza: " + id_lekarza);
         KartaPobytu kartaPobytu = entityManager.find(KartaPobytu.class, id_karty);
         Lekarz lekarz = entityManager.find(Lekarz.class, id_lekarza);
         if (kartaPobytu != null && lekarz != null && kartaPobytu.getLekarze().indexOf(lekarz) != -1) {
 
-            List<Recepta> recepty = kartaPobytu.getRecepty();
-            List<Diagnoza> diagnozy = kartaPobytu.getDiagnozy();
-            List<Operacja> operacje = kartaPobytu.getOperacje();
 
+            List<Diagnoza> diagnozy = new ArrayList<>(kartaPobytu.getDiagnozy());
+            List<Operacja> operacje = new ArrayList<>(kartaPobytu.getOperacje());
+            List<Recepta> recepty = new ArrayList<>(kartaPobytu.getRecepty());
+
+
+//            List<Diagnoza> diagnozy = entityManager.createQuery("from Diagnoza", Diagnoza.class).getResultList();
+            for (Diagnoza diagnoza : diagnozy) {
+                if (diagnoza.getId_lekarza() == id_lekarza && diagnoza.getId_karty() == id_karty) {
+                    entityManager.remove(diagnoza);
+                }
+            }
+//
+//            List<Recepta> recepty = entityManager.createQuery("from Recepta", Recepta.class).getResultList();
+            for(Recepta recepta: recepty){
+                if (recepta.getId_lekarza() == id_lekarza && recepta.getId_karty() == id_karty) {
+                    entityManager.remove(recepta);
+                }
+            }
+//
+//            List<Operacja> operacje = entityManager.createQuery("from Operacja", Operacja.class).getResultList();
             for(Operacja operacja: operacje){
-
-                if(operacja.getId_lekarza() == id_lekarza){
+                if (operacja.getId_lekarza() == id_lekarza && operacja.getId_karty() == id_karty) {
                     entityManager.remove(operacja);
                 }
             }
 
-            for(Recepta recepta: recepty){
-
-                if(recepta.getId_lekarza() == id_lekarza){
-                    entityManager.remove(recepta);
-                }
-            }
-
-            for(Diagnoza diagnoza: diagnozy){
-
-                if(diagnoza.getId_lekarza() == id_lekarza){
-                    entityManager.remove(diagnoza);
-                }
-            }
-
             kartaPobytu.removeLekarz(lekarz);
+            lekarz.removeKartaPobytu(kartaPobytu);
 
         }
     }
